Initial webapp-template to deploy static websites
This is just a template/example of a very simple static website.

(This will be deployed similarly to other [shiny apps](https://github.com/moj-analytical-services/rshiny-template))

When the docker container is built, the files under `www` will be copied in
the directory `/usr/share/nginx/html` and served by the web server (NGINX).

The `Dockerfile` is very simple and it's based on the [NGINX Docker image](https://hub.docker.com/r/library/nginx/).
I decided to use the `stable-alpine` for the following reasons:
 - use `latest` is very unpredictable. I thought use `stable` was a good
   tradeoff between "use the newest" and use a stable version which may
   include security updates.
 - as our users will be unlikely to think about updating NGINX, use `stable`
   will give them the latest version which is stable
 - use the `-alpine` variant is suggested (by the image creator) as this is
   smaller/safer and for this use case is probably more then plenty.

NOTE: The `EXPOSE 80` line is technically redundant by I decided to be
      explicit.